server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Application server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi[standard]>=0.121.0",
    "pyjwt>=2.10.1",
    "websockets>=15.0.1",
]
